+++ src/cppNGS/FileLocationProviderRemote.cpp
@@ -43,6 +43,12 @@ FileLocation FileLocationProviderRemote::getRepeatExpansionImage(QString locus) 
 	return getOneFileLocationByType(PathType::REPEAT_EXPANSION_IMAGE, locus);
 }
 
+FileLocationList FileLocationProviderRemote::getQcFiles() const
+{
+	//TODO GSvarServer
+	return FileLocationList();
+}
+
 FileLocationList FileLocationProviderRemote::getFileLocationsByType(PathType type, bool return_if_missing) const
 {
 	FileLocationList output;